+++ CHANGELOG.rst
@@ -4,4 +4,6 @@ Changelog
 0.1.0 - unreleased
 ------------------
 
+* Initial implementation of log analyzer and report object.
+
 * Initial package structure, docs, requirements, test scripts.